msm8974: added ml-listener to init.target.rc
This commit required for enable ml-listener to run
automatically on boot time.
The ml-listener calculates on-demands  hashes of requested files and
partitions. In order to be able to do that, it needs to run with
root privilege.
The calculated hashes are being saved in the SFS, and can only be read
by an authenticated request.


Change-Id: Ic0f5d99360d99baf21a2db05582060388f18d7f4